# feat(channel-web): more req per sec