--- a/modules/channel-web/src/views/lite/store/index.ts
+++ b/modules/channel-web/src/views/lite/store/index.ts
@@ -15,6 +15,7 @@ import {
   EventFeedback,
   Message,
   MessageWrapper,
+  QueuedMessage,
   StudioConnector
 } from '../typings'
 import { downloadFile, trackMessage } from '../utils'
@@ -67,6 +68,8 @@ class RootStore {
 
   @observable
   public botUILanguage: string = chosenLocale
+
+  public delayedMessages: QueuedMessage[] = []
 
   constructor({ fullscreen }) {
     this.composer = new ComposerStore(this)
@@ -126,8 +129,12 @@ class RootStore {
       return
     }
 
-    this.currentConversation.messages.push({ ...event, conversationId: +event.conversationId })
-    this.currentConversation.typingUntil = event.userId ? this.currentConversation.typingUntil : undefined
+    const message: Message = { ...event, conversationId: +event.conversationId }
+    if (this.isBotTyping.get() && !event.userId) {
+      this.delayedMessages.push({ message, showAt: this.currentConversation.typingUntil })
+    } else {
+      this.currentConversation.messages.push(message)
+    }
   }
 
   @action.bound
@@ -383,15 +390,18 @@ class RootStore {
     this.isBotTyping.set(true)
 
     this._typingInterval = setInterval(() => {
-      const typeUntil = this.currentConversation && this.currentConversation.typingUntil
-      if (!typeUntil || isBefore(new Date(typeUntil), new Date())) {
+      const typeUntil = new Date(this.currentConversation && this.currentConversation.typingUntil)
+      if (!typeUntil || isBefore(typeUntil, new Date())) {
         this._expireTyping()
+      } else {
+        this.emptyDelayedMessagesQueue(false)
       }
     }, 50)
   }
 
   @action.bound
   private _expireTyping() {
+    this.emptyDelayedMessagesQueue(true)
     this.isBotTyping.set(false)
     this.currentConversation.typingUntil = undefined
 
@@ -405,6 +415,19 @@ class RootStore {
       this.botUILanguage = lang
       localStorage.setItem('bp/channel-web/user-lang', lang)
     })
+  }
+
+  @action.bound
+  private emptyDelayedMessagesQueue(removeAll: boolean) {
+    while (this.delayedMessages.length) {
+      const message = this.delayedMessages[0]
+      if (removeAll || isBefore(message.showAt, new Date())) {
+        this.currentConversation.messages.push(message.message)
+        this.delayedMessages.shift()
+      } else {
+        break
+      }
+    }
   }
 
   /** Returns the current conversation ID, or the last one if it didn't expired. Otherwise, returns nothing. */