--- a/modules/channel-web/src/views/lite/typings.d.ts
+++ b/modules/channel-web/src/views/lite/typings.d.ts
@@ -251,6 +251,11 @@ export interface Message {
   timeInMs: number
 }
 
+export interface QueuedMessage {
+  message: Message
+  showAt: Date
+}
+
 export interface HTMLInputEvent extends Event {
   target: HTMLInputElement & EventTarget
 }