--- a/src/bp/core/repositories/users.ts
+++ b/src/bp/core/repositories/users.ts
@@ -3,6 +3,7 @@ import { DataRetentionService } from 'core/services/retention/service'
 import { inject, injectable } from 'inversify'
 import Knex from 'knex'
 import _ from 'lodash'
+import LRU from 'lru-cache'
 import ms from 'ms'
 
 import Database from '../database'
@@ -25,6 +26,7 @@ export class KnexUserRepository implements UserRepository {
 
   private batches: Dic<Row> = {}
   private flushLock: boolean = false
+  private userCache = new LRU({ max: 10000, maxAge: ms('1h') })
   private readonly flushInterval = ms('10s')
 
   constructor(
@@ -75,8 +77,17 @@ export class KnexUserRepository implements UserRepository {
     }
   }
 
+  private getCacheKey(channel: string, userId: string): string {
+    return `${channel}_${userId}`
+  }
+
   async getOrCreate(channel: string, id: string, botId?: string): Knex.GetOrCreateResult<User> {
     channel = channel.toLowerCase()
+    const cacheKey = this.getCacheKey(channel, id)
+
+    if (this.userCache.has(cacheKey)) {
+      return { result: <User>this.userCache.get(cacheKey), created: false }
+    }
 
     const ug = await this.database
       .knex(this.tableName)
@@ -104,6 +115,7 @@ export class KnexUserRepository implements UserRepository {
         otherChannels: []
       }
 
+      this.userCache.set(cacheKey, user)
       return { result: user, created: false }
     }
 
@@ -127,6 +139,7 @@ export class KnexUserRepository implements UserRepository {
         }
       })
 
+    this.userCache.set(cacheKey, newUser)
     return { result: newUser, created: true }
   }
 
@@ -155,6 +168,8 @@ export class KnexUserRepository implements UserRepository {
     }
 
     await req
+
+    this.userCache.del(this.getCacheKey(channel, user_id))
   }
 
   async updateAttributes(channel: string, user_id: string, attributes: any): Promise<void> {
@@ -167,6 +182,8 @@ export class KnexUserRepository implements UserRepository {
       .knex(this.tableName)
       .update({ attributes: this.database.knex.json.set({ ...originalAttributes, ...attributes }) })
       .where({ channel, user_id })
+
+    this.userCache.del(this.getCacheKey(channel, user_id))
   }
 
   private async _dataRetentionUpdate(channel: string, user_id: string, attributes: any) {