Removes some awaits, batches db calls and adds caching to users to allow more requests per seconds.

# Benchmarks

**http_req_duration** : stats for requests duration performing a test of 100 rps for 1 min
**http_reqs** : stats for requests per second in a test of 50 workers sending 5000 requests

## Without NDU & QNA

#### New
http_req_duration : avg=2.72ms   min=915.9µs  med=996.7µs  max=93.75ms
http_reqs : 5050    643.147186/s

#### Old
http_req_duration : avg=6.66ms   min=1.98ms   med=2.99ms   max=189.53ms p(90)=6.01ms   p(95)=12.96ms
http_reqs : 5050    195.00504/s

## With NDU & QNA

New
http_req_duration : avg=50.59ms  min=0s       med=36.9ms   max=568.59ms p(90)=104.71ms p(95)=151.55ms
http_reqs : 5050    656.746671/s

Old
http_req_duration : avg=344.6ms  min=2.98ms   med=345.04ms max=742.99ms p(90)=503.65ms p(95)=542.57ms
http_reqs : 5050    121.348228/s